NotificationMessage component prototype
Adding NotificationMessage component. At first I had a Notification
component that could render any "type" of notification, but changed
directions to have this component be presentational and render all
notifications. I have stubbed the idea of "notification types" and the
data for individual notifications in a `notification` directory. I will
look to replace this with logic that will live in the redux store and
probably be handled by a container component later.