transactions page: broken page on missing statement ids
It appears that it's possible to have a payload from the server with
statementIds referenced by transactions that are not in the list of statements.

Our code now gracefully deals with this by just omitting those statements when
looking up the invidual statements by their IDs.

Added a case to one of the fixtures for this to ensure that storybook loads
correctly in this instance.

Refactored the map and code in `getStatementsById` to more clearly break up its
steps so we can filter out nulls in the middle.