Notifications: Adding Notifications Widget
Adding the basic outline of the Notification Widget component. The
Widget is essentially a list of messages with some notion of them being
read or unread. This component is still purely functional and contains
no state, so the read/unread state of each message is passed in as well
as a function to marking messages read.

The Widget contains a link to a "Notification Center", which is a page
in an application containing notifications. Since the widget has no
knowledge of the application or its routing, this navigational component
must be passed in. The widget will render this component, but add a
className to it to encourage styling matching other items in the widget
footer.