Notifications: Aligning styles to Figma
It turns out that rem units are not well suited for use in this repo
because of the presence of Ant D, which seems to set the body font size
to 14px. We have also not decided a strategy of using a relative unit so
I'm dropping it for now and using pixels to ensure alignment with
designs in Figma.