fix styles for cells with links in sortable table
Regression of styles in sortable table and consequently in
statements table was introduced during statements page extraction.
Links in table has to be by default with black colored and change
color to blue on hover. Also font-size for text cells was incorrect,
12px instead of required 14px.

Current fix applies these styles on base elements with some cleanup
for styles.