refactor Pagination component
Previously, Pagination component has been created but not used
as a single version for pagination. In addition, Ant design components
were imported right into page views with duplicated styles and it brought
us into situation with multiple versions of almost identical functionality.

This change is a step forward to remove duplicated components and use single
version of pagination component. Along with this, `Pagination` component is
refactored with respect to latest designs.

`ResultsPerPageLabel` component is extracted as a standalone component which
displays number of results per paginated view. This functionality is used in
several places so it's a good time to refactor it as well.